qcacld-3.0: Update scan request IE's with default scan IE's
1) If g_qcn_ie_support is enabled, Save QCN IE along with default scan
   IE's received from Supplicant at init time.
2) Compare and update scan IE's with saved default scan IE's for each
   scan offload request to FW.

CRs-Fixed: 2033203

+++ core/hdd/inc/wlan_hdd_main.h
@@ -807,7 +807,7 @@ typedef struct hdd_scaninfo_s {
 	tSirAddie scanAddIE;
 
 	uint8_t *default_scan_ies;
-	uint8_t default_scan_ies_len;
+	uint16_t default_scan_ies_len;
 	/* Scan mode */
 	tSirScanType scan_mode;
